pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "cloudinary>=1.44.1",
    "fastapi>=0.135.1",
    "fastapi-mail>=1.6.2",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sendgrid>=6.12.5",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
]
